LPS-86009 Consistent typing.

+++ modules/apps/document-library/document-library-api/src/main/java/com/liferay/document/library/configuration/DLConfiguration.java
@@ -118,6 +118,6 @@ public interface DLConfiguration {
 		deflt = "0", description = "maximum-number-of-versions-help",
 		name = "maximum-number-of-versions", required = false
 	)
-	public long maximumNumberOfVersions();
+	public int maximumNumberOfVersions();
 
 }
